(PC-13531) [API] fix:criterion: split big migration
Split migration that added ondelete='CASCADE' to four FK into four
smaller migrations. This is needed to avoid locks/timeout errors.
