+++ api/src/pcapi/alembic/versions/20220222T083056_90c1a1eeeee0_add_index_on_booking_reimbursementdate.py
@@ -7,7 +7,7 @@ from pcapi import settings
 
 
 revision = "90c1a1eeeee0"
-down_revision = "ddedcf7e2c67"
+down_revision = "56c6d018551c"
 branch_labels = None
 depends_on = None
